packages/tools/preconf-profit-calc/requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
click>=8.0.0